solve the image update bug

+++ Controllers/UserController.js
@@ -10,6 +10,7 @@ module.exports = {
   UserRegisration: async (req, res, next) => {
     try {
       //   -- get Data from the body
+      console.log(req.body);
       const { firstName, lastName, phoneNumber, email, password } = req.body;
       if (!firstName) {
         return next(new ErrorHandler("Plaese Enter Your First Name", 400));
@@ -175,19 +176,28 @@ module.exports = {
       );
       if (req.file) {
         const userAavatar = User?.Avatar;
-        const filepath = path.join(__dirname, "../uploads", userAavatar);
-        fs.unlink(filepath, async (err) => {
-          if (err) {
-            console.log(`Error in file deleting ${err}`);
-            return res.status(400).json({ message: "Error in file deleting" });
-          } else {
-            console.log("file deleted successfuly");
-            const file = req.file.filename;
-            var fileUrl = path.join(file);
-            User.Avatar = fileUrl;
-            await User.save();
-          }
-        });
+        if (userAavatar) {
+          const filepath = path.join(__dirname, "../uploads", userAavatar);
+          fs.unlink(filepath, async (err) => {
+            if (err) {
+              console.log(`Error in file deleting ${err}`);
+              return res
+                .status(400)
+                .json({ message: "Error in file deleting" });
+            } else {
+              console.log("file deleted successfuly");
+              const file = req.file.filename;
+              var fileUrl = path.join(file);
+              User.Avatar = fileUrl;
+              await User.save();
+            }
+          });
+        } else {
+          const file = req.file.filename;
+          var fileUrl = path.join(file);
+          User.Avatar = fileUrl;
+          await User.save();
+        }
       }
       res.status(200).json({
         success: true,
